requirements.txt
numpy==2.2.4
scipy==1.15.2
opencv-python==4.11.0.86
open3d==0.19.0
trimesh==4.6.8
hydra-core==1.3.2
omegaconf==2.3.0
tqdm==4.67.1
pyyaml==6.0.2
matplotlib==3.10.1
huggingface-hub==1.4.1
d2nt==0.1.3